Rename actor setup package
